+++ src/main/java/domain/ladder/BridgeGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public interface BridgeGenerator {
     Bridge generate();
