+++ src/test/java/domain/LadderGameTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ladder.Bridge;
+import domain.ladder.BridgeGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
