+++ src/test/java/domain/ladder/LadderPositionsTest.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.ladder;
 
+import domain.ExceptionType;
+import domain.LadderGameException;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
