+++ src/test/java/domain/ladder/LadderTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
