+++ src/test/java/domain/ladder/RowGeneratorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
