+++ src/test/java/domain/ladder/WidthTest.java
@@ -1,7 +1,9 @@
-package domain;
+package domain.ladder;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.ExceptionType;
+import domain.LadderGameException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
